legacy/prototypes/requirements_serial.txt
PyQt6>=6.4.0
pyqtgraph>=0.13.0
pyserial>=3.5
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.5.0
